requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# AI
anthropic>=0.18.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
greenlet>=3.0.0

# WhatsApp/Twilio
twilio>=8.13.0

# Web Scraping
beautifulsoup4>=4.12.0
httpx>=0.26.0
lxml>=5.1.0

# Scheduling
apscheduler>=3.10.0

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Language Detection
langdetect>=1.0.9

# Utilities
python-multipart>=0.0.9
pytz>=2024.1
